Fix runner sharding key specs
These cannot be run on the schema version where the migration was
finalized because we already have a not null constraint on the sharding
key on that version.

This isn't ideal because this means the finalize migration only works if
all the records have already been migrated and it doesn't have to run
the background migration inline. I think this is fine since 17.5 was a
required stop and all record should've already been migrated by then.