Copy Sidekiq Web assets into Webpack output
The Sidekiq Web interface ships with CSS, images, and JavaScript files.
These files were previously only available on a container that installed
the Sidekiq gem, but the Workhorse container does not have any gems
installed. Copy these files into the `public/assets/sidekiq` directory
and point Sidekiq to use them.

This requires some changes to the Sidekiq gem to allow for assets
to be stored in a location outside of the gem.

Relates to https://gitlab.com/gitlab-com/gl-infra/production/-/issues/20862