Set security_project_tracked_context during vulnerability ingestion
We are working on tracking vulnerabilities across different branches. To
do this, `vulnerability_reads`, `vulnerability_occurrences`, and
`vulnerability_statistics` need to be associated with a
`security_project_tracked_context_id`. This change sets the ID during
vulnerability ingestion.

Continuous vulnerability scanning creates vulnerabilities from multiple
different projects / pipelines at the same time. So, we associate the
tracked context with a `finding_map` instance, and use
`TrackedContextFinder` to find or create the context. This class has a
cache based on pipeline ID so that we only execute queries once per
pipeline.