Add database triggers to update ci_pipeline_iids and ensure uniqueness
Adds database triggers that act when a ci_pipelines
record is inserted/deleted/updated (on iid) to update
and cleanup ci_pipeline_iids accordingly. This table
is meant to track all pipeline iids for a given project
so that it can ensure uniqueness across all partitions
on a database level.

Changelog: added