Fix ai_panel_spec under Vue 3
This fixes the underlying cause for the failure of the spec under Vue 3.
The next commit reverts the skipping of the spec under Vue 3.

The cause is interesting: my guess is that there is a Vue component name
collision. Both `new_chat_button.vue` and `navigation_rail.vue` had the
same `name` property. My guess is that caused something somewhere in
`@vue/test-utils@2` to blow up.

It might be this warrants a fix upstream. It's also possible that
`vue-test-utils-compat` is responsible. Further investigation is needed.