Smooth out observability access request flow
When a user was requesting observability access, they were redirected to
the setup page which was checking for the group's observability
settings. They aren't there because they're created in a background job.
This change adds a provisioning parameter to the redirect to the setup
page to indicate that the group is being provisioned. It won't fix
everything, but it will fix most of the issues.