Add role-based permissions to AI instance and namespace settings table
Currently we want to add an execute, manage and enable for projects
permission for agents and flows. These permission can be configured for
the instance or for top-level groups later on. Currently these
permissions are allowed to unset, as we don't want to invalid existing
setting entries.

Changelog: added
EE: true