Configurable tables for background migrations
Background migrations not always write to the table they iterate on. In
these cases it does not make sense to put the migration on hold because
of vacuum running for this table. This provides a way for migration to
explicitly provide list of tables to be checked for vacuum activity.