Unbox topics index page and remove faulty count
The list of topics is the only element on the page. No reason to put in
a box (called "crud container").

This change also revealed a bug: The `@topics.count` counted only the
elements of the current page, so the max was 20.

Also causing db queries in a template is a bad practise.

The paginated topics list disables counts in the controller for a good
reason: There can be thousands of different topics on one instance.

Changelog: fixed