Lock workflow in Redis when running workflow in workhorse
As described in https://gitlab.com/gitlab-org/gitlab/-/issues/548686
and https://gitlab.com/gitlab-org/gitlab/-/issues/580581 it's not good
that users can trigger the same workflow to be running twice
concurrently. Behaviour is undefined and likely bad for users.

This MR implements a simple locking mechanism using the existing Redis
distributed locking code we have in workhorse. When a
`StartWorkflowRequest` message is received in an open websocket
connection we attempt to acquire a lock for that workflowID in Redis. If
we cannot then we fail immediately. Otherwise we acquire the lock and
only release it once both the websocket and gRPC connections are closed.