Introduce namespace_id and namespace_path params
This MR updates the project import API endpoints to accept a target
namespace path or ID in explicitly named parameters, and marks the
original `namespace` parameter as deprecated. This provides clarity to
consumers of the API as to whether they should provide an ID (`123`) or
path (`widget-co/projects`). In turn, we at GitLab gain clarity on how
to use the input to find the namespace object.

The three updated endpoints are:

- `/api/v4/projects/import`
- `/api/v4/projects/remote-import`
- `/api/v4/projects/remote-import-s3`

The original behaviour is preserved, but if `namespace_path` or
`namespace_id` are provided, they will be used instead of `namespace`.

Changelog: changed