Updates logic on how foundational agents are available
Before, foundational agents were only available if the user had a
default namespace defined (or inferred) and that namespace enabled
foundational agents. However, that caused a bit of confusion on why they
would not be available on projects like gitlab if the user was part of
the gitlab project.

Thus, the new logic returns what is set for default namespace if it
exists. If not, it looks at the project the user is currently at, and
checks if that namespace has it enabled.

Changelog: changed
EE: true