Return early when no paths passed to blobs resolver
The blobs resolver includes a `paths` argument which is an `Array` of
`String` values and is `required`. In GraphQL `required: true` doesn't
ensure that the `Array` type argument contains any objects so we should
return early instead of executing the rest of the code when paths is
empty.

We already test that no records are returned when `paths` is empty so no
spec changes are required.