Use let_it_be for projects in slow banzai filter specs
This replaces a lot of the usages of `let` to create projects (and
groups) with `let_it_be` in some of the slower banzai filter specs.

Since we are usually not modifying the projects (and groups), we can
create them once with `let_it_be` and save about 0.2s to 0.6s per test
example.