Refactor and improve the performance of alive pipeline builds count
We have two similar methods doing almost the same calculation:

- jobs_count_in_alive_pipelines
- builds_count_in_alive_pipelines

The only difference is that the latter one only counts the builds;
it does not include trigger or other jobs.

Not only this is wrong but also not relevant the context they are used.

In this change, I'm consolidating the usage and add a small performance
gain by calling `jobs_count_in_alive_pipelines` only once.

The changes are behind the feature flag
`ci_refactor_jobs_count_in_alive_pipelines`.