doc: Clarify release asset link schemes
The current documentation claims, that "any URL"[1] is supported, which
is wrong: There are many schemes[2] but GitLab only allows `http`,
`https` and `ftp`. Using anything else like `docker` or `file` results
in an API error:

> release failed after 0.97 seconds. error=POST
> https://…/api/v4/projects/…/releases/…/assets/links:
> 400 {message: [Url is blocked: Only allowed schemes are http, https, ftp]}

The web-based GUI[3,4] and the backend[5] already validate the scheme,
but using the CLI[6] is it easy to run into that error as the constaint
is not documented.

Change the wording to make it clear, that only those 3 schemes are
allowed.

[1]: https://en.wikipedia.org/wiki/URL
[2]: https://www.iana.org/assignments/uri-schemes/uri-schemes.xhtml
[3]: https://gitlab.com/gitlab-community/gitlab-org/gitlab/-/blob/master/app/assets/javascripts/releases/components/asset_links_form.vue?ref_type=heads#L179
[4]: https://gitlab.com/gitlab-community/gitlab-org/gitlab/-/blob/master/app/assets/javascripts/releases/stores/modules/edit_new/getters.js?ref_type=heads#L97
[5]: https://gitlab.com/gitlab-community/gitlab-org/gitlab/-/blob/master/app/models/releases/link.rb?ref_type=heads#L14
[6]: https://docs.gitlab.com/ci/yaml/#releaseassetslinks

Signed-off-by: Philipp Hahn <p.hahn@avm.de>