Refactor last processed index to message ID in duo agentic chat
Now UiChatLog for a chat workflow returns a message id. We can
use it to determine which messages have already been rendered.

Once this change in place, we can also signal DWS that we no
longer need the messages that have already been rendered, so
the server doesn't have to send them anymore.

Changelog: added