Quarantine a flaky test
The `` Security::UnassignPolicyConfigurationsForExpiredLicensesCronWorker#perform when on self-managed when runtime limit is reached reschedules the worker with the correct cursor ``
test matches one of the following conditions:
1. has either ~"flakiness::1" or ~"flakiness::2" label set, which means the number of reported failures
is at or above 95 percentile, indicating unusually high failure count.

2. has ~"severity::1" label set, which means the number of reported failures
[spiked and exceeded its daily threshold](https://gitlab.com/gitlab-org/ruby/gems/gitlab_quality-test_tooling/-/blob/c9bc10536b1f8d2d4a03c3e0b6099a40fe67ad26/lib/gitlab_quality/test_tooling/report/concerns/issue_reports.rb#L51).

This MR quarantines the test. This is a discussion starting point to let the
responsible group know about the flakiness so that they can take action:

- accept the merge request and schedule the associated issue to improve the test
- close the merge request in favor of another merge request to delete the test

Please follow the
[Flaky tests management process](https://handbook.gitlab.com/handbook/engineering/testing/flaky-tests/#flaky-tests-management-process)
to help us increase `master` stability.

Please let us know your feedback in the
[Engineering Productivity issue tracker](https://gitlab.com/gitlab-org/quality/engineering-productivity/team/-/issues).

Related to https://gitlab.com/gitlab-org/quality/test-failure-issues/-/issues/16561.

This change was generated by
[gitlab-housekeeper](https://gitlab.com/gitlab-org/gitlab/-/tree/master/gems/gitlab-housekeeper)
in [CI](https://gitlab.com/gitlab-org/quality/engineering-productivity/team/-/jobs/12250236354) using the `Keeps::QuarantineFlakyTests` keep.

To provide feedback on your experience with `gitlab-housekeeper` please create an issue with the
label ~"GitLab Housekeeper" and consider pinging the author of this keep.

Changelog: other