Ensure we only mock once for `begin_update`
It is discovered that the hook can be called multiple times when we're
only saving once. The reason why it's called multiple times is that some
associations can cause the parent to be saved again, calling the hook
again. That's a reasonable behaviour, but here we only want to hook
once.