Add user_id param to secondary email confirmation links
To prepare for de-scoping `emails.confirmation_token` to be per-user
rather than globally unique, we need to ensure the controller has the
`email.user_id` to look up the appropriate email to confirm.

This change ensures the confirmation link only for secondary emails
(using the Email model) includes the `user_id` parameter. This will be
used by the confirmations controller to look up the Email record to
confirm, once the new index is in place.

Changelog: changed