Add configurationForGroup field
Following the pattern of configurationForProject. This change also fixes
an issue with the existing configurationForProject field, where due to
the key being scoped to only the project_id, the same item consumer
would be returned for the project, even for different items. It adds the
item ID into the key also, so they're correctly returned.

Part of:

https://gitlab.com/gitlab-org/gitlab/-/issues/580381