Add flag to bypass tag commit checks during tag hooks
This flag allows us to experiment with the new tag commit check when
deciding if a tag should generate a pipeline or not.

This is required to test a report that this change isn't working as
intended. We are currently working on the assumption that this isn't
working because the dereferenced_commit is nil but we haven't been able
to confirm this in the logs.

This flag will allow us to turn the feature of for specific projects and
validate that it is this change that is causing problems.