Make pipeline configuration review optional
This makes the reviewer requirement for pipeline configuration files
optional and opens up approvals to all maintainers.

This reduces the load on pipeline maintainers and shortens the MR cycle
time because an MR that touches the pipeline config no longer requires a
review by a pipeline maintainer.

By keeping the optional pipeline configuration reviewers limited to the
pipeline maintainers, we keep it easy for contributors to request a
pipeline configuration review from domain experts at their discretion.