Log workflow_id on lock failure
These logs statements relate to work done in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/212532 .
See that MR for more information.

This MR simply ensures the "Failed to acquire workflow lock" log
statement is accompanied by a workflow_id so it's easier to correlate
the failures when users provide us with a workflow ID. This was an
oversight in the first MR because clearly we can see that this is indeed
logged for success cases.