Read epic data from the WorkItem in the REST API
This changes the way how we read data for epics in the REST API, but
reading directly from the work item, instead of the epic object.

Since the work item is the source of truth as most services write to the
work item first, we should start reading from it in our APIs.

This does not change the API because the data is bi-directionally synced.