Speed up pipeline-tier-N jobs
This speeds up the `pipeline-tier-N` jobs (such as `pipeline-tier-3`) by
setting the `needs` jobs to an empty array, reducing the artifact
download time from around 23s to 0s because we don't need the
`clone-gitlab-repo` artifact.

This is only a minor improvement, which saves us less than 100 compute
hours per month. But it's still neat to see the job complete faster.