Use correct service account for flows
When executing a flow with a trigger, the flow is now executing using
the service account belonging to the flow.

A composite identity is also used for triggering the flow, as well as
for the OAuth token that is used by the pipeline.

The Duo Developer user is also not added to the project automatically
anymore.