Fix order of admin mode omniauth log
When a user re-authenticates for admin mode, we should create audit and
authentication events after the admin flow has completed, not before it
has been initiated. When the logs are created before the admin flow is
complete, a log entry appears in the UI even when the authentication
fails, which is confusing.