Remove get_issue from MCP config
It was added for testing purposes and now when we have other
tools in the list, we can remove it.

Also, get_issue is replaced in DWS by work items tools, so we
don't plan to migrate a DWS get_issue tool to an MCP get_issue
tool, so it's another reason to remove it from this list.