Use same apolloProvider everywhere
The legacy issues list has been updated to the work items list
and no longer supports caching. The issues dashboard list will
be updated to the work items list and won't support caching
at that point too. So to simplify the code, we'll use the same
apolloProvider everywhere, which also reduces the bug surface area.