Rename check_health to server_available?
- Renamed SecretsManagerClient#check_health method to server_available?
following ruby style guide. It returns true/false

- We have a SecretsManager::Helpers module but the
ErrorResponseHelper was NOT inside this module.
Moved it inside the correct module.

- The secrets_manager_inactive_response method inside ErrorResponseHelper
module mentioned above is also included in GroupSecretsManager
but the method specifically deals with project_secrets_manager.
Updated the error message to be a generic one that works for both
projects and groups.