Pass in isLoading arg to getCurrentModel function
Fixes a bug where a user's model selection in the agentic chat panel is being cleared pre-maturely from localStorage.

The getCurrentModel() function should return null when availableModels is still loading, but we weren't passing in the loading state.
We need to return early when the query is sitll loading otherwise the getSavedModel() function will run and attempt to validate the model
against an empty availableModels array. It will fail and trigger the removal of the saved model from localStorage.

Changelog: fixed
EE: true