Drop use of GetTreeEntries() RPC in Secret Push Protection
This commit removes the usage of `GetTreeEntries()` RPC from Secret Push
Protection in favour of using `FindChangedPaths()` solely that now
includes the commit SHA for each changed path entry.

In order to do so, we build out a map to look up the changed paths
including their associated commit SHAs for each individual payload, and
while building out the message to display to the user, we pull the
commit SHA and the file path directly from that map instead of calling
`GetTreeEntries()` for each commit.

EE: true