Drop SidekiqWebStatic middleware
This middleware is no longer needed now that we copy Sidekiq assets into
the Webpack-generated asset bundle in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/213303.

Previously we needed this middleware because the Webservice pod used the
X-Sendfile interface to have Workhorse send the assets. The assets were
not present on the Workhorse container because they previously only
existed if the Sidekiq gem were installed. Now that the assets are now
available, we can remove this middleware and simplify the stack.

Relates to
https://gitlab.com/gitlab-com/gl-infra/production/-/issues/20862

Changelog: changed