Enable Duo runner restrictions feature flag by default
This is related to
https://gitlab.com/gitlab-org/gitlab/-/issues/577521#note_2912564213

This feature was introduced in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/213431 and
instructions for testing are described in there.

In short it limits Duo flows to only running on instance or top level
group runners.

This MR enables the feature flag by default but still offers
self-managed customers the option to disable it in case it's a problem
for them.