Publish event for transfers earlier
The transfer service could timeout after the transaction. This timeout
leads to the GroupTransferedEvent not being emitted.

This leads into a semi-atomic transfer operation, because we
have listeners on this events and we rely on it being triggered.

Ideally all other (long-running) data migrations should only be handled
in workers, instead of the TransferService directly.

To mitigate the problem for now, we publish the GroupTransferedEvent
immediately after updating the namespaces.

Changelog: fixed