Put Job Router discovery endpoint behind feature flag
This change set puts the new job router discovery endpoint behind a
feature flag to better control the roll out of the feature flag.

There already exists a feature flag on the Runner, but we need a way to
disable the Job Router entirely on an instance level.

There are two feature flags:

- `job_router_instance_runners`: enable job router for instance runners only.
- `job_router`: enable job router for top-level groups or globally.

Refs https://gitlab.com/gitlab-org/gitlab/-/issues/583449

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/583473