Make tooling review optional
This makes the reviewer requirement for tooling files and scripts
optional and opens up approvals to all maintainers.

This reduces the load on tooling maintainers and shortens the MR cycle
time because an MR that touches tooling files and scripts no longer
requires a review by a tooling maintainer.

By keeping the optional tooling reviewers limited to the original
tooling maintainers, we keep it easy for contributors to request a
tooling review from domain experts at their discretion.