Link composite identity when assigning users
When checking abilities for composite identities we always return false
if the composite identity is not linked.

This causes issues when assigning a service account user to an issue or
merge request, or when setting it as a reviewer.

In these cases we should link the identity before checking the
privileges of the user.