Closing the 2FA banner remove query parameter
In we made the 2FA setup banner to appear in the 2FA page. We forgot to
`initClose2faSuccessMessage` in the 2FA page.

This functions removes the `two_factor_auth_enabled_successfully=true`
query parameter from the URL,
https://gdk.test:3443/-/profile/two_factor_auth?two_factor_auth_enabled_successfully=true,
when the banner is closed.

I think the idea is to persist the banner until it is acknowledged
(closed) by the customer. Personally, I feel the implementation is a bit
questionable, but for consistency with the account page, let's keep it.