Raise Error on 4XX and 5XX responses from Openbao
Prior to this change, we were NOT raising errors
on 4XX and 5XX status codes from Openbao. As a result of this,
we noticed a bug scenario in production where Openbao server was down
but the requests to Openbao in provisioning a secrets manager did not
fail causing the feature to be in an unusable state

This change fixes the issue by using raise_error
middleware of Faraday. We explicitly allow 404
status code.

References
https://gitlab.com/gitlab-org/gitlab/-/work_items/582364
https://lostisland.github.io/faraday/#/middleware/included/raising-errors