Fix build queue removal during database replication lag
Prevent builds from being removed from the queue when the replica
database has not caught up with the primary. Previously, builds could be
incorrectly removed when they appeared non-pending on the replica but
were still pending on the primary, causing them to be lost.

Relates to https://gitlab.com/gitlab-com/request-for-help/-/issues/3690

Changelog: fixed