Remove Duo Core access to Duo Classic Chat
This is done by checking the new UP duo_classic_chat,
which is baesd from duo_chat UP, but without Duo Core.

Feature flag is used to determine which UP to use.

After this, the chat modes, feature and UP mapping are:

|         | feature      | unit primitive   |
|---------|--------------|------------------|
| Classic | duo_chat     | duo_classic_chat |
| Agentic | agentic_chat | duo_chat         |

This reverses earlier changes, where agentic chat was mapped to duo_agent_platform feature.
Now it is remapped to agentic_chat feature instead.