Add gitlab--duo tag to external agent CI pipelines
This is related to
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/213431 . In that
MR we introduced a requirement at
https://gitlab.com/gitlab-org/gitlab/-/blob/ebe7f776864aa6fffafa0b3e455b90cd4909e854/ee/app/services/ee/ci/register_job_service.rb#L20
that all jobs with source `duo_workflow` must be run on instance or
top-level group runners. This means that any jobs that are end up on a
subgroup or project runner will fail with an error. But the only
sensible way we can ensure that jobs don't end up on subgroup or project
runners is by giving them some tag (we chose `gitlab--duo`) which allows
runner admins to only include that tag on top-level or instance runners.

But in that MR we missed the fact that these "flow trigger" or "external
agents" are also being marked as `duo_workflow` source so the
restriction will apply to them too. But they don't actually have any
`gitlab--duo` tag so they will just be randomly assigned a runner. This
will cause problems for users because they can't control where the job
ends up and sometimes it will end up on a runner that is not allowed to
run this type of job.

This MR simply ensures that these jobs also get the `gitlab--duo`
tag.