Enable Job Token to retrieve Tag via REST API
This change set adds support for the CI Job Token as a valid
authentication token to retrieve project tags via the REST API. The
listing endpoint for tags already supports it and so, I think the
endpoint to retrieve a single tag should support it too.

This enables full support for the GitLab CLI to auto-login to create a
release of which part is to check for the release tag. Without this
feature we have to resort to page through the listing endpoint to find
our tag which is not particularly good.

Changelog: changed