Usage Quotas: Setup repository health section
This change is part of multiple
MRs behind the feature flag project_repositories_health_ui
focused on implementing a repostiory health dashboard
on the Usage Quotas page.

This change is focused on hooking up the repository
health section component as well as introducing
the API call required to get the respected data.