Add tables_to_check_for_vacuum to MRDF partition BBM
The BBM was merged in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/205280, before
this directive was added. Since the partitioning backfill is projected
to be completed in _months_, we're hoping that adding this now will have
some measurable performance improvements, shortening the total run time.

Normally, we don' modify running migrations, however I confirmed with
@alexives that this was an acceptable course of action because we're not
modify the actual _action behavior_ of the migration, and we don't want
to sacrifice the ~5 weeks of progress we've already made. This change,
once merged, should be picked up by all subsequently run jobs in the
batch.