gitaly: Add Gitaly version requirement in contributing doc
Following a lengthy discussion about how to handle deploying
Gitaly and its clients independently during a GitLab upgrade,
it has been agreed that Gitaly's clients should not use features
that are going to be deployed in the same version as the clients.

In other words, Gitaly clients should always use the previous
version of Gitaly.

This commit adds the doc to reflect this policy.

References:

https://gitlab.com/gitlab-org/gitaly/-/issues/7028