Limit count of merge requests on the list view
Limits the count in the tabs on the merge request list to improve
performance. Currently we will load the full count of merge requests for
the project or group which can get expensive with more merge requests.

This change limits the number to show `999+` if there are more than 999
merge requests in the project or group.

https://gitlab.com/gitlab-org/gitlab/-/issues/563367