Backfill organization_id on pool_repositories
This commit provides the batched background migration to backfill
organization_id on pool_repositories.

Additionally it provides the following:

- database trigger to set organization_id to cover scenarios where
  model hooks are skipped
- concurrent index on organization_id
- foreign key constraint (NOT VALID) on organization_id
- NOT NULL (NOT VALID) constraint on organization_id
- specs

The trigger was causing test cases to fail where we want to test that
there is no organization_id, because it was getting added on INSERT.

This commit temporarily drops the trigger for those tests and then addds
it back.

Changelog: added