Move request body params extraction logic to own class
GrapeOpenapi::Converters::RequestBodyConverter is a big class that has
many different responsibilities, one of which is to extract the body
parameters from a request.

Since the class is quite big and complex and the body parameter
extraction logic is well self-contained, this commit moves it to its
own dedicated class: GrapeOpenapi::Models::RequestBody::Parameters, and
references it in GrapeOpenapi::Converters::RequestBodyConverter.