Move request body param schema generaiton logic to own class
GrapeOpenapi::Converters::RequestBodyConverter is a big class that has
many different responsibilities, one of which is to generate the schema
of parameters in the request body.

Since the class is quite big and complex and the body parameter schema
generation logic is well self-contained, this commit moves it to its
own dedicated class: GrapeOpenapi::Models::RequestBody::ParameterSchema
and references it in GrapeOpenapi::Converters::RequestBodyConverter.