Add migration to backfill occurrence_id to representation information
This MR backfills vulnerability.finding_id to
vulnerability_representation_information.vulnerability_occurrence_id.
This will allow us to change the relationship of the table to
occurrences instead of vulnerabilities. This will support the tracking
of vulnerabilities on multiple branches.

We began setting this column for new records in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/214287

Changelog: changed