Implement internal job update endpoint with failure message support
This change set introduces a new internal REST API endpoint for the Job
Router to update a job state. With that it also introduces a new
`failure_message` field in addition to the existing `failure_reason`
field and a new supported `job_router_failure` value that can be used
with the new `failure_message` field.

This allows the Job Router to provide more useful error messages to the
user when it has to fail a job. For example, when admission control
fails we want to attach the reason provided by that controller or things
like with controller failed the job.

Refs https://gitlab.com/gitlab-org/gitlab/-/issues/584552